+++ src/main/webapp/app/graph/graph.component.html
@@ -14,5 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License. 
 -->
 
-<google-chart #chart id="chart" [type]="type" [columns] = "columns" [data]="graphData" [options]="options">
+<google-chart #chart id="chart" [type]="type" [columns]="columns" [data]="graphData" [options]="options">
 </google-chart>
